backend/requirements.txt
fastapi
uvicorn
crewai
google-ortools
pydantic
python-dotenv
langchain-google-genai  # For Gemini integration